NOLF2 Object.lto: changed most enums to enum class
using enum class (c++11) to be able to forward declare enums.

+++ NOLF2/ObjectDLL/ObjectShared/ServerUtilities.h
@@ -36,7 +36,7 @@ class CPlayerObj;
 #define LOAD_BOOL(variable) variable = (pMsg->Readbool() ? LTTRUE : LTFALSE);
 #define SAVE_bool(variable) pMsg->Writebool(variable);
 #define LOAD_bool(variable) variable = pMsg->Readbool();
-#define SAVE_DWORD(variable) pMsg->Writeuint32(variable);
+#define SAVE_DWORD(variable) pMsg->Writeuint32((uint32)variable);
 #define LOAD_DWORD(variable) variable = pMsg->Readuint32();
 #define LOAD_DWORD_CAST(variable, clazz) variable = (clazz)pMsg->Readuint32();
 #define SAVE_QWORD(variable) pMsg->Writeuint64( variable );
